Ignore selinux when recursing over /home/$user
When re-running the manifest, the /home directory may already be
populated with lots of files. In those cases, recursing over
home directories takes a lot of time, due to the file resource
calling SELinux on every file to get its context. In some cases, it
was even causing OOM errors.