Do not run chown on worker's home directory if unneeded
By default, we're currently running a recursive chown in the worker's
home directories. This can be extremely time-consuming if the home
directory is populated, so we'll change it to only run if the home
owner does not match.

Please note this opens a possibility: if the first chown times out,
the worker's home will probably be set, but some sub-directories may
not. An admin will need to correct this manually.

This commit fixes https://github.com/rdo-infra/puppet-dlrn/issues/14