Add SSH key for review.rdoproject.org if setting up gerrit
When gerrit reviews are set up for a worker, the first time a review
is started SSH will ask the user to confirm the host key for the
gerrit server, in our case review.rdoproject.org. Since this requires
manual input, the DLRN process will hang.

Solving this by adding the SSH host key to each user that configures
gerrit.

This fixes #20