Added support for passive server
Currently, dlrn infra in RDO project is formed by two servers a
primary and a passive one. This commit adds proper support for both
servers to use the same hiera config file by creating a new parameter
server_type which can be set to primary (default) or passive.

Even when enabled in workers hiera configuration, passive servers will not enable:
- cron jobs for run-dlrn.sh
- rsync of new builds (they are suposed to be the destination of the synchronization)
- Opening reviews in gerrit automatically on FTBFS
- Sending mails on build failure
Additionally, passive servers will configure apache to redirect to buildlogs when users
trying to access to current-passed-ci or current-tripleo

As part of this commit, apache configuration is fully integrated in puppet-dlrn and
no more ssl_setup.sh is required. In case certificates are not installed, parameter
enable_https con be set to false.