Rename centos-ocata user to centos-master-uc
When introducing the master worker using upper-constraints, we
used centos-<branch> as name. This creates issues when branching a
new stable branch, since we need to re-create the user and create
a new one, breaking continuity for the centos-master repos.

This change renames the centos-ocata user to centos-master-uc. With
this, the fugure centos-ocata branch can be created from scratch,
and centos-master-uc can maintain continuity.

Change-Id: I3c09e7fd4490125c71c1f56e6a70f4123fe934c0