Add run-purge script
Just like we have a run-dlrn.sh helper script, add a run-purge.sh
one to simplify purge operations on workers. It provides a set
of sane defaults (60 days retention, do not remove anything on the
special symlinked directories), sends output to a log file and makes
sure no other dlrn commands can be run at the same time by using
the same lock file.

This script can be used in the future in a cron job if needed.

Change-Id: I0331f0b49d830e83726960158301dae715e65f9c