Prepare for stable/ocata branch
Upstream is about to start creating the stable/ocata branch for
libraries, so it's time to prepare the stable Ocata worker.

Note that the URL and symlinks used by the stable/ocata worker are
temporary, since centos-master-uc is already using the centos7-ocata
symlinks. Once the worker starts to be used, we'll switch the
URLs and symlinks to it.

Change-Id: I1d8c743e2cbb5717efb4daf29ba01edf120c7b51