When purging, only run rsync once a week
The process of running rsync to the backup instance after a full
purge can be very slow on long-lived workers (we've seen >6h), due
to rsync scanning the whole repos directory. This is an issue,
because during the rsync process we still cannot build packages.

Let's run rsync only once a week, when the package build load is low.

Change-Id: I626b3a3d2bc7d5a47bad51672646ce392876a392