Prepare for stable/pike
We are getting ready to bootstrap the stable/pike worker, so preparing
the required puppet-dlrn bits.

Note that the URL and symlinks used by the stable/pike worker are
temporary, since centos-master-uc is already using the centos7-pike
symlinks. Once the worker is bootstrapped, we'll switch the URLs and
symlinks to it.

Also, to speed up bootstrap, we are setting mock_tmpfs_enable to
true. We may set it to false later, if needed.

Change-Id: I690c5a296559522acd0eded8e317954757608cb5
Story: 168
Task: 411