Additional configuration to support SELinux=Enforcing
Some additional SELinux booleans were needed to support the DLRN API,
together with a small policy to allow httpd-based process to run
rsync when importing a repository and a specific fcontext for the
worker's home directories.

While at it, we are moving to the latest puppet-selinux from [1], to
reduce our old module usage.

[1] - https://github.com/voxpupuli/puppet-selinux

Change-Id: I605b9670b37eb291cc70e6950e48801471a81506