Stop fetching the home page with wget
It doesn't make sense to have a repository just to maintain a static
html page. Let's move it to puppet-dlrn so we save some needless
complexity around using wget, setting up a cron and so forth.

We can eventually even template the html file if we want to in order
to generate the list of workers or API endpoints dynamically.

Change-Id: I3300c0018392f9f7c60cfa8e9124c6637af9800f