run-purge: do not remove symlinks created only on remote
The current rsync command-line used for run-purge.sh synchronized
all contents in the base repo directory /home/${USER}/data/repos.
This means that, if we create a symlink only on the remote node, it
will be removed as part of the rsync process.

By adding the * wildcard to the command line, we ensure that any
file existing on the target directory but not on the source directory
will be preserved. This can only be an issue if some file had to be
removed, but in that case we should do it in both locations.

Change-Id: Id8b2be269e256dae1882a94641fb4ac0b331c7f3