Add coreutils to install list in Fedora Rawhide mock config
There are two mutually-exclusive packages in Fedora: coreutils and
coreutils-single. For some reason, when coreutils is installed by
yum as a dependency (for binutils, for example), coreutils-single
is installed. Later in the process, coreutils tries to be installed,
causing a conflict and an error to build the package. This only
happens to yum and not dnf.

We can workaround this by explicitly installing coreutils. It is
a transitive dependency to gcc, whichi is already installed, so
we are not increasing the number of packages installed in the chroot.

Change-Id: I622f975119b21d9ff544ac9933ce1ed51a4339e6