Sync symlinks from public to internal instance before purging
When running run-purge.sh, we are synchronizing all symlinks from
/home/$USER/data/repos to the public instance.

We have changed the way promotions are done to use the DLRN API, and
this is only run in the public instance now. This creates a potential
issue where an older symlink (in the internal instance) may overwrite
a newer one on the public instance.

To fix this, let's sync all symlinks from the public instance before
purging.

As an additional fix, add the current-tripleo-rdo-internal symlink
to the list of protected symlinks during the purge.

Change-Id: I5d0a66bec178e989d670354a7d389128c14d8c82