Prepare for stable/queens
We are getting ready to bootstrap the stable/queens worker, so let's
prepare the required puppet-dlrn bits.

Note that the URL and symlinks used by the stable/queens worker are
temporary, since centos-master-uc is already using the centos7-queens
symlinks. Once the worker is bootstrapped, we'll switch the URLs and
symlinks to it.

Change-Id: I246c0b9e706585efa5b28704d217f2c383a2acaf