Prepare for stable/rocky
We are getting ready to bootstrap the stable/rocky worker, so let's
prepare the required puppet-dlrn bits.

Note that the URL and symlinks used by the stable/rocky worker are
temporary, since centos-master-uc is already using the centos7-rocky
symlinks. Once the worker is bootstrapped, we'll switch the URLs and
symlinks to it.

Change-Id: Ic16f0e3cd91b915dad4143282e280cbc6dbd6f9f