Use createrepo instead of createrepo_c in update-deps.sh
We've detected that, when using --retain-old-md in createrepo_c
it doesn't limit the number of retained metadata and keeps all
old metadata, making repodata directory to keep growing on each
run. As the performed tests, createrepo seems to work fine and
although time to run is a bit longer, it's still aceptable (just
seconds to run on centos-master-uc).

Change-Id: I863c28f528ac9c6d8eb481cf1d114202e967a2f7