Remove centos7-virt-container repo from delorean-deps files
In [1], container related packages have been added to deps.yml so they
will be treated as deps in RDO packages instead of consume them from
VirtSIG repo.

This patch removes VirtSIG repo from the delorean-deps.repo files as it
may introduce updates or conflicts with the packages in the RDO repos.

Change-Id: I35067735ad69dc1bea7929185b890792d41029b2