Temporarily manage the SSL renewal cron job manually
The current letsencrypt module we are using does not allow us to
set a specific hour, minute or weekday for the renewal cron job,
which can cause some issues if it triggers the daily httpd reload
while some operations are running.

The best option will be to update to the latest puppet-letsencrypt,
but that will require some time and testing. In the meantime, let's
manage the cron job manually and make sure it is only executed
once every week.

Change-Id: Id8b95c401ea4a9e45997333b22060c68954c97b8