Prepare for stable/stein
We are getting ready to bootstrap the stable/stein worker, so let's
prepare the required puppet-dlrn bits.

Note that the URL and symlinks used by the stable/stein worker are
temporary, since centos-master-uc is already using the centos7-stein
symlinks. Once the worker is bootstrapped, we'll switch the URLs and
symlinks to it.

Also note we are creating a fedora-stein worker, in addition to a
centos-stein one.

Change-Id: Ie2d48a29a2b53d4a8869dc1d9ee104c149dcf72b