Prepare for stable/train worker
We are getting ready to bootstrap the stable/train worker, so let's
prepare the required puppet-dlrn bits.

Note that the URL and symlinks used by the stable/train worker are
temporary. Once the worker is bootstrapped, we'll switch the URLs and
symlinks.

Also note that we are not creating a fedora-train worker, since we
have plans of removing the fedora workers during the cycle, once
CentOS 8 is ready.

Change-Id: I9a8010fd0ed6d4ae2257b2dc3d6dadeea3847351