Ensure cron jobs are removed when set to false
Previously, when one of the variables controlling cron jobs for
a worker was not set, the cron job was simply ignored. This causes
issues when we want to disable a previously running cron job, because
it is not actually disabled.

This commit changes this, and makes sure the cron jobs are wiped
when not needed.

Change-Id: I99271ccbdda7b8817b924d1b86b2ee18f42a9abf